chore: 매칭 상태 enum 변수 분리
