+++ walkmong/walkmong/Global/Components/MatchingState.swift
@@ -1,0 +1,15 @@
+//
+//  MatchingState.swift
+//  walkmong
+//
+//  Created by 황채웅 on 12/20/24.
+//
+
+import Foundation
+
+enum MatchingState: String{
+    case matching = "매칭중"
+    case confirmed = "매칭확정"
+    case ended = "산책완료"
+    case cancelled = "매칭취소"
+}
