+++ walkmong/walkmong/Presentation/Walktalk/Chat/Views/WalktalkChatView.swift
@@ -139,9 +139,3 @@ class WalktalkChatView: UIView {
     }
 }
 
-enum MatchingState: String{
-    case matching = "매칭중"
-    case confirmed = "매칭확정"
-    case ended = "산책완료"
-    case cancelled = "매칭취소"
-}
